Route protection + optimization

+++ app/controllers/AvailabilityController.java
@@ -25,6 +25,10 @@ import java.util.concurrent.TimeUnit;
 public class AvailabilityController extends Controller {
 
     public Result createAvailability() {
+        String userRole = UserController.getCurrentRole();
+        if(!userRole.equals("Coach") || !userRole.equals("Admin")){
+            return unauthorized();
+        }
    /* Get user object from request */
         JsonNode json = request().body().asJson();
         /* Get user from json request */
@@ -217,6 +221,10 @@ public class AvailabilityController extends Controller {
     }
 
     public Result removeAvailability() {
+        String userRole = UserController.getCurrentRole();
+        if(!userRole.equals("Coach") || !userRole.equals("Admin")){
+            return unauthorized();
+        }
         JsonNode json = request().body().asJson();
         String availabilityId = json.findPath("availabilityId").asText();
         AvailabilityDB.removeAvailability(availabilityId);
